Added getPast events endpoint. Added validation to signup/callout/cancel functions to ensure they can only be called for events that (a) exist and (b) are in the future.

+++ controllers/attendance.js
@@ -41,38 +41,61 @@ const signup = (req, res, connection) => {
                             res.status(400).send('Bad request')
                         } else {
 
-                            // check if a attendance record exists already
-                            connection.execute('SELECT * FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                            // confirm event exists
+                            connection.execute('SELECT * FROM `events` WHERE event_id = ?', [event_id], (err, results, fields) => {
                                 if (err) {
                                     console.error(err)
                                     res.status(500).send('Server error')
                                 } else {
-                                    if (results.length > 0) {
-
-                                        // update sign up
-                                        connection.execute('UPDATE `attendance` SET signed_up = NOW(), called_out = NULL, character_id = ?, role_id = ?, tentative = ?, late = ?, note = ? WHERE event_id = ? AND discord_user_id = ?', [character_id, role_id, tentative, late, note, event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
-                                            if (err) {
-                                                console.error(err)
-                                                res.status(500).send('Server error')
-                                            } else {
-                                                res.status(200).send('Success')
-                                            }
-                                        })
-
+                            
+                                    // if length is zero then the event doesn't exist
+                                    if (results.length === 0) {
+                                        res.status(400).send('Bad request')
                                     } else {
 
-                                        // insert sign up
-                                        connection.execute('INSERT INTO `attendance` (event_id, discord_user_id, signed_up, called_out, character_id, role_id, tentative, late, note) VALUES (?, ?, NOW(), NULL, ?, ?, ?, ?, ?)', [event_id, jwt_data.body.discord_user_id, character_id, role_id, tentative, late, note], (err, results, fields) => {
-                                            if (err) {
-                                                console.error(err)
-                                                res.status(500).send('Server error')
-                                            } else {
-                                                res.status(200).send('Success')
-                                            }
-                                        })
+                                        // confirm event start date/time is still in the future
+                                        const startDate = new Date(results[0].start)
+                                        const now = new Date()
+                                        if (startDate < now) {
+                                            res.status(400).send('Bad request')
+                                        } else {
+
+                                            // check if a attendance record exists already
+                                            connection.execute('SELECT * FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                                                if (err) {
+                                                    console.error(err)
+                                                    res.status(500).send('Server error')
+                                                } else {
+                                                    if (results.length > 0) {
+
+                                                        // update sign up
+                                                        connection.execute('UPDATE `attendance` SET signed_up = NOW(), called_out = NULL, character_id = ?, role_id = ?, tentative = ?, late = ?, note = ? WHERE event_id = ? AND discord_user_id = ?', [character_id, role_id, tentative, late, note, event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                                                            if (err) {
+                                                                console.error(err)
+                                                                res.status(500).send('Server error')
+                                                            } else {
+                                                                res.status(200).send('Success')
+                                                            }
+                                                        })
+
+                                                    } else {
+
+                                                        // insert sign up
+                                                        connection.execute('INSERT INTO `attendance` (event_id, discord_user_id, signed_up, called_out, character_id, role_id, tentative, late, note) VALUES (?, ?, NOW(), NULL, ?, ?, ?, ?, ?)', [event_id, jwt_data.body.discord_user_id, character_id, role_id, tentative, late, note], (err, results, fields) => {
+                                                            if (err) {
+                                                                console.error(err)
+                                                                res.status(500).send('Server error')
+                                                            } else {
+                                                                res.status(200).send('Success')
+                                                            }
+                                                        })
+                                                    }
+                                                }
+                                            })
+                                        }
                                     }
                                 }
-                            })
+                            }) 
                         }
                     }
                 })
@@ -98,35 +121,58 @@ const callout = (req, res, connection) => {
                 res.status(400).send('Invalid token')
             } else {
 
-                // check if a attendance record exists already
-                connection.execute('SELECT * FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                // confirm event exists
+                connection.execute('SELECT * FROM `events` WHERE event_id = ?', [event_id], (err, results, fields) => {
                     if (err) {
                         console.error(err)
                         res.status(500).send('Server error')
                     } else {
-                        if (results.length > 0) {
 
-                            // update call out
-                            connection.execute('UPDATE `attendance` SET signed_up = NULL, called_out = NOW(), character_id = NULL, role_id = NULL, tentative = FALSE, late = FALSE, note = NULL WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
-                                if (err) {
-                                    console.error(err)
-                                    res.status(500).send('Server error')
-                                } else {
-                                    res.status(200).send('Success')
-                                }
-                            })
-
+                        // if length is zero then the event doesn't exist
+                        if (results.length === 0) {
+                            res.status(400).send('Bad request')
                         } else {
 
-                            // insert call out
-                            connection.execute('INSERT INTO `attendance` (event_id, discord_user_id, signed_up, called_out, character_id, role_id, tentative, late, note) VALUES (?, ?, NULL, NOW(), NULL, NULL, FALSE, FALSE, NULL)', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
-                                if (err) {
-                                    console.error(err)
-                                    res.status(500).send('Server error')
-                                } else {
-                                    res.status(200).send('Success')
-                                }
-                            })
+                            // confirm event start date/time is still in the future
+                            const startDate = new Date(results[0].start)
+                            const now = new Date()
+                            if (startDate < now) {
+                                res.status(400).send('Bad request')
+                            } else {
+
+                                // check if a attendance record exists already
+                                connection.execute('SELECT * FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                                    if (err) {
+                                        console.error(err)
+                                        res.status(500).send('Server error')
+                                    } else {
+                                        if (results.length > 0) {
+
+                                            // update call out
+                                            connection.execute('UPDATE `attendance` SET signed_up = NULL, called_out = NOW(), character_id = NULL, role_id = NULL, tentative = FALSE, late = FALSE, note = NULL WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                                                if (err) {
+                                                    console.error(err)
+                                                    res.status(500).send('Server error')
+                                                } else {
+                                                    res.status(200).send('Success')
+                                                }
+                                            })
+
+                                        } else {
+
+                                            // insert call out
+                                            connection.execute('INSERT INTO `attendance` (event_id, discord_user_id, signed_up, called_out, character_id, role_id, tentative, late, note) VALUES (?, ?, NULL, NOW(), NULL, NULL, FALSE, FALSE, NULL)', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                                                if (err) {
+                                                    console.error(err)
+                                                    res.status(500).send('Server error')
+                                                } else {
+                                                    res.status(200).send('Success')
+                                                }
+                                            })
+                                        }
+                                    }
+                                })
+                            }
                         }
                     }
                 })
@@ -153,22 +199,45 @@ const cancel = (req, res, connection) => {
                 res.status(400).send('Invalid token')
             } else {
 
-                // check if a attendance record exists
-                connection.execute('SELECT * FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                // confirm event exists
+                connection.execute('SELECT * FROM `events` WHERE event_id = ?', [event_id], (err, results, fields) => {
                     if (err) {
                         console.error(err)
                         res.status(500).send('Server error')
                     } else {
-                        
-                        // delete attendance record
-                        connection.execute('DELETE FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
-                            if (err) {
-                                console.error(err)
-                                res.status(500).send('Server error')
+
+                        // if length is zero then the event doesn't exist
+                        if (results.length === 0) {
+                            res.status(400).send('Bad request')
+                        } else {
+
+                            // confirm event start date/time is still in the future
+                            const startDate = new Date(results[0].start)
+                            const now = new Date()
+                            if (startDate < now) {
+                                res.status(400).send('Bad request')
                             } else {
-                                res.status(200).send('Success')
+
+                                // check if a attendance record exists
+                                connection.execute('SELECT * FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                                    if (err) {
+                                        console.error(err)
+                                        res.status(500).send('Server error')
+                                    } else {
+                                        
+                                        // delete attendance record
+                                        connection.execute('DELETE FROM `attendance` WHERE event_id = ? AND discord_user_id = ?', [event_id, jwt_data.body.discord_user_id], (err, results, fields) => {
+                                            if (err) {
+                                                console.error(err)
+                                                res.status(500).send('Server error')
+                                            } else {
+                                                res.status(200).send('Success')
+                                            }
+                                        })
+                                    }
+                                })
                             }
-                        })
+                        }
                     }
                 })
             }
